refactor(rust): Split compression.rs into modular structure
